Cumae Display: frame management.
Most of the work to produce this change was about coming up with a
trivial way to generate correctly aligned frame data.

Introducing: Cumae Framegen :)

Framegen is a Python script that generates correctly aligned C data
structures ready to be integrated into a Cumae app.

The script has already been integrated in the Cumae build system
(basically few neat Makefiles:)) in such a way that it will generate
frame data for whatever PBM file it finds in the gfx/ directory.

A C header file will be generated for each PBM file found in the gfx/
directory. The include path for such files will be <frames/FILENAME.h>.