Cumae Display: better context management.
This change brings along quite a bit of refactoring of the Cumae layer.
Most of the changes are non-functional as they are a refinement of
existing code.

A display context has been introduced with a better and more flexible
way to initialize the display. This should make supporting new display
variants a little bit easier.

One of the difficulties that have been found is about the delay function
that has been 'hacked' to be usable at runtime. It'll need to be
reviewed as it might waste quite a lot of cycles for simple delays.

Supporting new displays should be rather easy now; to do that:

1. Add the relevant information in the cumae/header/config.h
2. Fix the cm_display_type_e enum to include the new display
3. Provide the boilerplate code in cm_display_get_default_context
4. Remove the display check throughout the code to include the new type.
5. Test and fix
6. Fix and test
7. Test and test
8. And test again