LED mgmt, low-power mode and new GFX data.
This squashed commit brings along some long-needed change about the way
the status LED is handled, along with a simple - but very effective -
low power mode that kicks in after around 60 seconds from the last user
activity.

In order to achieve that, the prescaler has been set to /64 and given
that the clock is currently set to 12MHz, the Timer1 overflows every 3~
seconds.

When that happens, a "stand-by" frame is pushed on the G2 so that users
can enjoy a nice QR code that points to this repo.:)