New Cumae features: Permanent Data and UUID.
A simple permanent data struct has been introduced in the base Cumae
code. This is mainly meant to be used to differentiate devices.
It's also gonna be used as starting point for some sort of "preference"
system that I'm thinking to implement.

The first use case is with UUIDs: there are a couple of interesting new
Makefile targets. Issuing "make unique" will:

1. Initially create a new empty binary file
2. The binfile will then be filled with the relevant:
    a. Cumae version information
    b. a UUID
3. After that, the binfile will be burnt on the device EEPROM

Cumae will read the content of the EEPROM and will show the relevant
information at boot time.

The struct cm_permanent_data_s will be extensible.;)